requirements.txt
requests>=2.32.0
python-dotenv>=1.0.0
pyyaml>=6.0
duckdb>=1.1.0
pytest>=8.0.0
